requirements.txt
torch==2.3.0
torchvision==0.18.0
pillow==10.2.0
matplotlib==3.8.2
tqdm==4.66.1